requirements.txt
streamlit>=1.28.0
google-generativeai
pillow>=9.0.0
python-dotenv>=0.19.0
